This commit factors out most of analysis.py program.  Instead of always collecing C files in the subdirectory tree in order to reconcile it with Kmax's result, that part of the analysis is factored out into FileAnalysis.analyze, which can be enabled with the -F flag.
It also makes treating config options as Booleans the default, instead adding an additional -T flag to force Kmax to use the (experimental) tristate model.

The driver has been updated to also make using Booleans the default and also adds the -F flags to perform the analysis of the subdirectory for comparison against the Kmax results.