set src variable and account for both Kbuild and Makefile
some Kbuild source files use the $(src) variable, which is set to the
kbuild path.

also, some subdirectories have both a Makefile and a Kbuild Makefile,
so kmaxall will now add the Makefile to the list of pending
directories both are there.  currently kmax will look for Kbuild
first, then Makefile.  this should probably be made simpler and more
explicit in the future.