factor out reading in kconfig types
now reading in the kconfig_extract only happens once and is passed to
any call to print_model_as_config, which uses it to generate the
.config variable setting appropriately.

this also fixes a bug in the handling of no kconfig_extract file,
although this functionality may be unnecessary.  if removed, the
--kclause-formulas option needs to be paired with a --kconfig-extract
option as well.