count builds including unit as PASS even if there's a compiler error
previously, test_klocalizer.sh would fail when make.cross fails.  but
this would wrongly report a FAIL in cases where the selected
compilation unit is included in the build, but either there is a
compile-time error in building or some other compilation unit causes a
build failure.  all that test_klocalizer is counting is whether
klocalizer's configuration actually tried to compile the requested
compilation unit.