Complete rewrite of expression converter using python 3's ast module
The expression converter takes an expression extracted from Kconfig
and builds a z3 expression from it.  Previously, it used python 2's
built-in support for parsing and compilation of python to parse and
visit the expression tree.  python 3 provides a new module called ast
that has similar functionality, but a different api.