Revise the data structure holding kmax results
When running kmax with the -a flag, it will output the contents of
more Makefile variables other than obj-y, etc, i.e., the object files
not destined for the kernel binary.  The data structure is now a
dictionary from type of variable to contents.  Presence conditions are
not included in this output, which are still available in the default
output (-z).