Revised the API between the kextract C extension on python scripts
Now the python script, kextract, is a thin wrapper around the C
extensions (located in kextracts/).  The wrapper just provides an
option to select between the different versions of the extensions.

kextractlinux provides canned arguments to kextract, taking only the
architecture and output file from the user, as the originally-named
kextract script did.