kclause: introduce composite format for klocalizer
Currently, kclause files are dictionaries, which are needed to be parsed
by klocalizer on each run such that each clause in the dictionary is
accumulated to a constraints list. This process has a great share on
the overall running time the kclause by almost doubling it.

Create a new kclause format, i.e., composite kclause, which takes away
the need for traversing a dictionary, and instead stores the composite
clauses already accumuluated in a list.

kmax/klocalizer now takes --use-composite-kclause-formulas-files flag,
which tells klocalizer to interpret the kclause formulas file in composite
format.

scripts/kclause_to_composite.py can be used to convert kclause files into
composite kclause files. The functionality of this script can be later
embedded into the formula extractor/generator to have this seamless.