klocalizer: allow trying multiple archs without specifying a compilation unit
klocalizer is more than localizing a configuration for a compilation unit
or an architecture. For example, one might want to find a satisfying
configuration when an additional set of constraints are defined using:

klocalizer --all --constraints-file-smt2=path/to/constraints.smt

This currently fails with error message "ERROR: Trying multiple
architectures without a target compilation unit is not provided. Give one
architecture or --kclause-formula to generate a config file without
providing a compilation unit.".

Do not fail for such cases.

In addition, if a compilation unit is not provided, the command outputted
for building is not correct.

Fix the build command for cases without compilation units.