kextractor: refactor choice related code
kextractor assumes multiple prompts are not used as it is illegal in
kconfig, but it is possible to break its output. Instead, print a warning
when encountered with multiple prompts and only use the first prompt.

Some comments and commented code blocks are no longer used since fixes
were applied or the commented codes were replaced. Remove such blocks
to clean the code.