Automatically d/l cached formulas; version-specific .kmax folders
Firstly, the default directory holding formulas is now Linux version
specific.  For example, the kclause formulas will be at
.kmax/next-20210512/kclause

This is useful to avoid using the wrong formula cache when checking
out another version of Linux in a git repository.  klocalizer will
first check the version with git describe or make kernelversion if it
is not a git clone.  The consequence is that all formula caches will
remain on disk until manually deleted by the user.

Since the tool detects the version, it can tell whether the formulas
are already cached.  If not, it will try to find prebuilt formulas at
https://configtools.org/klocalizer/cache/VERSION.  If the user has a
folder for that version, however, i.e., if they already generated some
of the formulas, then no download will occur (even if not all the
architectures' formulas have been generated by the user yet).