Prevent adding a variable entry that has a False presence condition
In some cases, e.g., arch/mips/Makefile, a presence condition for a
branch will be False.  But if there is a variable assignment in that
branch, the symbol table may receive a definition that is only valid
under False, which is never possible.  This causes the variable to be
treated as defined, which in turn causes an exception when using the
variable in another conditional expression.  Instead, we filter out
any attempts to add infeasible entries so that the variable will be
treated as never having been defined in the first place.