klocalizer: add linux top makefile constraints
The top Linux kernel makefile is not passed to kmax since it is not a
Kbuild makefile. This makefile has inclusion constraints for some
directories like net/ (CONFIG_NET) and samples/ (CONFIG_SAMPLES).

Hard-code this in klocalizer to capture directory constraints from top
makefile.

These constraints in top makefile exist only in recent kernel versions
(> v5). However, the fix always adds the constraints for those
directories. It is assumed that doing so will not adversely affect the
constraints.