klocalizer: hash build system files to identify formulas id
Using git tags or kernelversion does not work in several edge cases.

Instead, hash the build system files, and use it to identify the build
system version.  Also use this build system id as the version id for
formulas.

The downside is that computing the hashes take some time (about 4
seconds on SSD).