klocalizer: allow specifying unit lines to localize
klocalizer was taking a list of compilation units to compile.

Now that the line-precision is being implemented, allow klocalizer to
take a list of lines to compile along with the units.

As before, a unit name is stil acceptable input. On top of that, a list
of lines can be provided. The list contains a list of integers and/or
ranges. A snippet from the code that shows how the input is interpreted:

    == Input ==                    == Output (unit, lines)==
    "kernel/fork.o"                ("kernel/fork.o", [])
    "kernel/fork.c"                ("kernel/fork.o", []) ('.c' -> '.o')
    "kernel/fork.o:[]"             ("kernel/fork.o", [])
    "kernel/fork.o:[5,7,10]"       ("kernel/fork.o", [5,7,10])
    "kernel/fork.o:[10-10]"        ("kernel/fork.o", [10])
    "kernel/fork.o:[5,7-10]"       ("kernel/fork.o", [5,7,8,9,10])
    "kernel/fork.o:[10-7]"         Prints error and terminates
    "kernel/fork.o:[10,"           Prints error and terminates