klocalizer: allow multiple approximate (repair) .config files
klocalizer used to allow a single approximate argument.

The upcoming mutual exclusion implementation will allow multiple .config
files to be outputted that cover different parts of the constraints.

To allow approximating to different .config files on different
architectures, let klocalizer take multiple approximate arguments.

If a config file is prefixed with "arch_name:", the config is used for
approximating for that architecture only. If no prefix is provided, it
is used as default for any architecture.

For example:

For example, "--approximate x86_64:1.config --approximate arm:2.config
--approximate others.config" will use 1.config for x86_64, 2.config for
arm, and others.config for any other architecture.