klocalizer: relax unit/srcfile extension check
Local methods unit2srcfile() and srcfile2unit() is used to change
extensions between .c and .o.  In some cases, they are called even if
the extension is already the target one (e.g., unit2srcfile() is called
while extension is already .c).

Change the assertions inside those methods to accept this, so that
they can be used with more flexibility, without needing to keep track of
what the current extension was.