arch: raise specific exception for Kconfig files that can't be opened
There is a specific case of kextract failure, which is when Kconfig
files cannot be opened, possibly due to missing Kconfig files. This is
a bug in the input Kconfig specifications rather than a kextract issue.

Therefore, raise a specific exception (Arch.CantOpenKconfigFiles) when
this happens, which is a subclass of
Arch.KextractFormulaGenerationError.