Use Kconfig default values when non-Boolean options get enabled
Previously, klocalizer would use fixed non-Boolean values, e.g., 0 for
numbers, and a prior update added support for reusing the values from
an input configuration when using the --repair option.  In this
update, if there is no setting from the input configuration file,
klocalizer will look at the default values in the Kconfig file (via
kextract), and use one of them for the value, falling back to the
fixed one if that isn't available.  Currently it picks the last
default value, since Kconfig allows for multiple defaults.  But in the
future, it would be best to use constraint solving to pick the right
default, since each default is associated with a condition when there
are multiple.