Remove assertions that require a config to repair
Previously, when generating a configuration file given constraints on
config options, files, and/or lines, an assertion would trigger
because there is no configuration file given as input to repair.
klocalizer should be able to generate a config file according to
constraints without an input file as was prior behavior.  This removes
the asserts allow configuration file generation to proceed.