Replace the optimized analysis loop with a multiprocessor pool
The body of the loop has been factored out into do_optimized, which
takes a triple of the elements from the original triple loop.  It
returns a tuple of the results that were originally collected within
the loop.

The creation of the pool can fail if ulimit is too low, so there is a
try-except to check for it.

The pool is run with imap_unordered, which goes over the triples as an
iterable in any order and runs do_optimized on each element.  The
results are run through tqdm to provide a progress bar.  The results
are looped over and saved in the unmet_constraints results structure.

The counter for optimized pass found cases has been moved into the
multiprocessor loop to avoid having to make a multiprocessor counter.