Substitute tristate test Booleans to support tristate value tests
When a kextract produces a comparison of a tristate option to y or m,
the expression converter replaces the comparison with a special option
that represents the comparison, e.g., tristate_m_CONFIG_SOUNDWIRE.

Only options that have such comparisons undergo the substitution.
Since most options don't, this reduces the additional constraints
needed for tristate options.

To implement the semantics of the tristate value comparisons in terms
of the configuration options variables, the following clause is added
to the configuration options' dependency expression,

CONFIG_A <-> tristate_y_CONFIG_A xor tristate_m_CONFIG_A

i.e., option A is enabled if and only if it is set to either y or m
but not both.

Currently, the above clause is added to every expression that tests a
tristate against y or m.  But this can happen in multiple dependency
expressions, so it will add redundant clauses in that case.