Add additional constraints for tristate values
This adds a constraint that connects the Boolean configuration option
with its potential tristate values, i.e.,

CONFIG_A <-> CONFIG_A==y ^ CONFIG_A==m

This means that whenever CONFIG_A is enabled, one of the special predicate
Booleans, CONFIG_A==y or CONFIG_A==m, are enabled as well (and
vice-versa).  This allows for the use of tristate-value specific
dependencies to work correctly.  <-> is biimpliciation and ^ is xor.

This also adds a constraint that forces select to choose y, which
seems to be the right behavior, though perhaps their are cases where
the selected option can only be m.